Extract stadium forecast logic into apply_stadium_action.rs
Moves the Mesagoza/Fragrant Forest/Area Zero/Kid's Room UseStadium
forecasting functions out of the already-large apply_action.rs into
their own module, following the same apply_*_action.rs pattern used
for attacks and trainer cards.
